Add category specific RSS feeds
This was because Chris Shaw asked. It wasn't particularly painful,
although it took me an hour or so to get this out of the door.

There's a basic feed generator plugin, a feed layout file, and links to
the RSS feeds on category pages and the category index.

It's pretty straightforward, but I should probably take the time to
publish the plugin and document this on my site - I'm sure this will
bite me in the butt if I don't.