Add read-only Disqus comments from an export
Now the only way to engage with my site is through email, although I'll
probably add webmention support at some point in the future.

I used Gemini 2.5 Pro to put together the bones of the plugin, HTML, and
CSS changes. This saved me a good couple of hours and having to relearn
Ruby.